src/requirements.txt
fastapi
mangum
pydantic
google-genai
httpx
psycopg
psycopg-binary